Changed controller to task
