--- src/controller.h
@@ -1,7 +1,0 @@
-/** @file   controller.h
- * 
- *  @author Joshua Hoye
- *  @date   11/30/2021
- */
- 
-void compute (void* p_params);
