+++ src/task_controller.h
@@ -1,0 +1,7 @@
+/** @file   controller.h
+ * 
+ *  @author Joshua Hoye
+ *  @date   11/30/2021
+ */
+ 
+void compute (void* p_params);
